rename map to widget_map
