+++ met-api/src/met_api/resources/widget_map.py
@@ -21,7 +21,7 @@ from flask_restx import Namespace, Resource
 from met_api.exceptions.business_exception import BusinessException
 from met_api.auth import jwt as _jwt
 from met_api.services.widget_map_service import WidgetMapService
-from met_api.schemas.map import WidgetMapSchema
+from met_api.schemas.widget_map import WidgetMapSchema
 from met_api.utils.roles import Role
 from met_api.utils.util import allowedorigins, cors_preflight
