+++ met-api/src/met_api/services/widget_map_service.py
@@ -2,7 +2,7 @@
 from http import HTTPStatus
 from datetime import datetime
 from met_api.exceptions.business_exception import BusinessException
-from met_api.models.map import WidgetMap as WidgetMapModel
+from met_api.models.widget_map import WidgetMap as WidgetMapModel
 
 class WidgetMapService:
     """Widget Map management service."""
